redesign and name correction
